+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-08-10 03:27:08
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 6d600380f35146a20951a218fa0f761b89b13ebf
+
+## [ForwardWidgets-2kuai] - 2025-08-10 03:27:08
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: ccd2fcb882820b11a536da7195b2ef9013f6a8eb
+
+## [ForwardWidgets-pack1r] - 2025-08-10 03:27:08
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 9419b152a431ccd0f8ade0704a9844c174a70277
+
+
 ## [ForwardWidgets-huangxd] - 2025-08-09 03:13:21
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main